--- a/lib/cli.js
+++ b/lib/cli.js
@@ -1,7 +1,6 @@
 /* @flow */
 "use strict";
 
-const chalk = require("chalk");
 const dynamicRequire = require("./dynamicRequire");
 const EOL = require("os").EOL;
 const getModulePath = require("./utils/getModulePath");
@@ -11,6 +10,7 @@ const needlessDisablesStringFormatter = require("./formatters/needlessDisablesSt
 const path = require("path");
 const resolveFrom = require("resolve-from");
 const standalone = require("./standalone");
+const tc = require("turbocolor");
 
 /*:: type meowOptionsType = {
   argv?: string[],
@@ -502,9 +502,9 @@ module.exports = (argv /*: string[]*/) /*: Promise<void>|void*/ => {
             const foundWarnings = linted.maxWarningsExceeded.foundWarnings;
 
             process.stdout.write(
-              chalk.red(`Max warnings exceeded: `) +
+              tc.red(`Max warnings exceeded: `) +
                 `${foundWarnings} found. ` +
-                chalk.dim(`${maxWarnings} allowed${EOL}${EOL}`)
+                tc.dim(`${maxWarnings} allowed${EOL}${EOL}`)
             );
             process.exitCode = 2;
           }