--- a/lib/formatters/needlessDisablesStringFormatter.js
+++ b/lib/formatters/needlessDisablesStringFormatter.js
@@ -1,7 +1,7 @@
 "use strict";
 
-const chalk = require("chalk");
 const path = require("path");
+const tc = require("turbocolor");
 
 function logFrom(fromValue) {
   if (fromValue.charAt(0) === "<") return fromValue;
@@ -19,7 +19,7 @@ module.exports = function(report) {
       return;
     }
     output += "\n";
-    output += chalk.underline(logFrom(sourceReport.source)) + "\n";
+    output += tc.underline(logFrom(sourceReport.source)) + "\n";
     sourceReport.ranges.forEach(range => {
       output += `start: ${range.start}`;
       if (range.end !== undefined) {