--- a/lib/formatters/stringFormatter.js
+++ b/lib/formatters/stringFormatter.js
@@ -1,11 +1,11 @@
 "use strict";
 
 const _ = require("lodash");
-const chalk = require("chalk");
 const path = require("path");
 const stringWidth = require("string-width");
 const symbols = require("log-symbols");
 const table = require("table");
+const tc = require("turbocolor");
 const utils = require("postcss-reporter/lib/util");
 
 const MARGIN_WIDTHS = 9;
@@ -25,11 +25,11 @@ function deprecationsFormatter(results) {
   }
 
   return uniqueDeprecationWarnings.reduce((output, warning) => {
-    output += chalk.yellow("Deprecation Warning: ");
+    output += tc.yellow("Deprecation Warning: ");
     output += warning.text;
     if (warning.reference) {
-      output += chalk.dim(" See: ");
-      output += chalk.dim.underline(warning.reference);
+      output += tc.dim(" See: ");
+      output += tc.dim.underline(warning.reference);
     }
     return output + "\n";
   }, "\n");
@@ -42,7 +42,7 @@ function invalidOptionsFormatter(results) {
   const uniqueInvalidOptionWarnings = _.uniq(allInvalidOptionWarnings);
 
   return uniqueInvalidOptionWarnings.reduce((output, warning) => {
-    output += chalk.red("Invalid Option: ");
+    output += tc.red("Invalid Option: ");
     output += warning;
     return output + "\n";
   }, "\n");
@@ -102,7 +102,7 @@ function formatter(messages, source) {
   let output = "\n";
 
   if (source) {
-    output += chalk.underline(logFrom(source)) + "\n";
+    output += tc.underline(logFrom(source)) + "\n";
   }
 
   const cleanedMessages = orderedMessages.map(message => {
@@ -112,7 +112,7 @@ function formatter(messages, source) {
       location.line || "",
       location.column || "",
       symbols[severity]
-        ? chalk[levelColors[severity]](symbols[severity])
+        ? tc[levelColors[severity]](symbols[severity])
         : severity,
       message.text
         // Remove all control characters (newline, tab and etc)
@@ -122,7 +122,7 @@ function formatter(messages, source) {
           new RegExp(_.escapeRegExp("(" + message.rule + ")") + "$"),
           ""
         ),
-      chalk.dim(message.rule || "")
+      tc.dim(message.rule || "")
     ];
 
     calculateWidths(row);
@@ -148,7 +148,7 @@ function formatter(messages, source) {
     })
     .split("\n")
     .map(el =>
-      el.replace(/(\d+)\s+(\d+)/, (m, p1, p2) => chalk.dim(p1 + ":" + p2))
+      el.replace(/(\d+)\s+(\d+)/, (m, p1, p2) => tc.dim(p1 + ":" + p2))
     )
     .join("\n");