--- a/lib/formatters/verboseFormatter.js
+++ b/lib/formatters/verboseFormatter.js
@@ -1,8 +1,8 @@
 "use strict";
 
 const _ = require("lodash");
-const chalk = require("chalk");
 const stringFormatter = require("./stringFormatter");
+const tc = require("turbocolor");
 
 module.exports = function(results) {
   let output = stringFormatter(results);
@@ -16,7 +16,7 @@ module.exports = function(results) {
   const checkedDisplay = ignoredCount
     ? `${results.length - ignoredCount} of ${results.length}`
     : results.length;
-  output += chalk.underline(`${checkedDisplay} ${sourceWord} checked\n`);
+  output += tc.underline(`${checkedDisplay} ${sourceWord} checked\n`);
   results.forEach(result => {
     let formatting = "green";
     if (result.errored) {
@@ -30,20 +30,20 @@ module.exports = function(results) {
     if (result.ignored) {
       sourceText += " (ignored)";
     }
-    output += _.get(chalk, formatting)(` ${sourceText}\n`);
+    output += _.get(tc, formatting)(` ${sourceText}\n`);
   });
 
   const warnings = _.flatten(results.map(r => r.warnings));
   const warningsBySeverity = _.groupBy(warnings, "severity");
   const problemWord = warnings.length === 1 ? "problem" : "problems";
 
-  output += chalk.underline(`\n${warnings.length} ${problemWord} found\n`);
+  output += tc.underline(`\n${warnings.length} ${problemWord} found\n`);
 
   _.forOwn(warningsBySeverity, (warningList, severityLevel) => {
     const warningsByRule = _.groupBy(warningList, "rule");
     output += ` severity level "${severityLevel}": ${warningList.length}\n`;
     _.forOwn(warningsByRule, (list, rule) => {
-      output += chalk.dim(`  ${rule}: ${list.length}\n`);
+      output += tc.dim(`  ${rule}: ${list.length}\n`);
     });
   });