**Which issue, if any, is this issue related to?**

- None, as it's a miscellaneous perf improvement.

**Is there anything in the PR that needs further explanation?**

- No, it's self explanatory.

---

Hi! This PR swaps chalk for [turbocolor](https://github.com/jorgebucaran/turbocolor) that is ligther (no deps) and has ~20x faster load time and runtime. This is a non-breaking change. This is effectively a "drop-in" replacement.

Let me know if this looks good to you and if you'd like to accept my contribution. 👋 

###### turbocolor/chalk

```
# Load Time
chalk: 15.190ms
turbocolor: 0.777ms

# All Colors
chalk × 8,729 ops/sec
turbocolor × 158,383 ops/sec

# Chained Colors
chalk × 1,838 ops/sec
turbocolor × 39,830 ops/sec

# Nested Colors
chalk × 4,049 ops/sec
turbocolor × 59,833 ops/sec
```