@gucong3000 2018-07-21
> This is a non-breaking change.

We need to do more checking:

- [ ] Compatibility with CI
  - [ ] appveyor
  - [ ] gitlab-ci
  - [ ] travis-ci
  - [ ] circleci
  - [ ] jenkins
- [ ] Compatibility with CLI arg: `--no-color` & `--color`
- [ ] Compatibility with [supports-color](https://github.com/chalk/supports-color#readme)