@jeddy3 2018-07-21
I suggest we leave it with `chalk` at the moment.

We have a lot of moving parts in stylelint at the moment, and swapping out `chalk` will add another variable into the mix.